+++ src/lib/knownSlugs.js
@@ -1,1 +1,0 @@
-export { getFilesystemSlugs } from '../data/prototypes';